Restructure base-detail-view: move hook and config-toolbar into subfolders
