+++ assets/studio/js/src/components/base-detail-view/hooks/use-detail-view.ts
